funcion de busqueda de ubicaciones

+++ zaraga.weather/Models/WeatherLocation.cs
@@ -5,8 +5,8 @@ namespace zaraga.weather.Models
     public class WeatherLocation
     {
         public int id { get; set; }
-        public decimal lat { get; set; }
-        public decimal lon { get; set; }
+        public double lat { get; set; }
+        public double lon { get; set; }
         public string name { get; set; }
         public string region { get; set; }
         public string country { get; set; }
